Added DAO layer, implemented factory design pattern and added new features to procedure call
Implemented a DAO layer and a factory to support different databases. UI needs to be updated to input driverClassName. Difference class is a static class now. Better suits the needs for line-by-line and all at once comparison. There is no need to store an object for the class.  Updated the list view to show which parameters were used at each comparison. Removed the "current parameters" label.